Disabled scrollbar and added scaling

+++ frontend/src/board/arrow.rs
@@ -40,7 +40,6 @@ impl PartialEq for Arrow {
 }
 impl Eq for Arrow {}
 
-#[derive(Hash)]
 pub struct ControlPoint {
     pub point: Coords,
     pub vector: Coords,
@@ -54,8 +53,8 @@ fn display_coords_poly(coords: &Coords) -> String {
     format!("{},{}", coords.x, coords.y)
 }
 
-const TR_DX: i32 = 7;
-const TR_DY: i32 = (TR_DX as f32 * 0.5774) as i32;
+const TR_DX: f64 = 7.0;
+const TR_DY: f64 = TR_DX * 0.5774;
 
 fn triangle_html(cp: &ControlPoint) -> Html {
     // TODO: this triangle ending works only with horizontal and long enough 'end' control vector
